Calculate a slightly more robust estimate of growth rate and proportion based on a single model at a range of

Calculate a slightly more robust estimate of growth rate and proportion based on a single model at a range of

pointProportionEstimate(
  simpleTimeseries,
  dates = max(simpleTimeseries$date) - 3,
  window = 14,
  weekly = "weekday",
  includeModel = TRUE,
  ...
)

Arguments

simpleTimeseries
- the timeseries containing date, value and total

dates
- dates at which to evaluate the model

window
- the window of data

weekly
- either "weekday","weekend" or "none" to define whether to fit a fixed effect model to the weekday, or the is.weekend

includeModel
- keep the fitted model as a list column for fit analysis

...

Value

a dataframe of evaluation dates, growth rates, proportions and model fit