Calculate a reproduction number estimate using the Wallinga 2007 estimation using empirical generation time distribution. This uses resampling to transmit uncertainty in growth rate estimates

Calculate a reproduction number estimate using the Wallinga 2007 estimation using empirical generation time distribution. This uses resampling to transmit uncertainty in growth rate estimates

rtFromGrowthRate(
  simpleTimeseries,
  infectivityProfile,
  yMatrix = infectivityProfile$yMatrix,
  aVector = infectivityProfile$aVector,
  bootstraps = 20 * dim(yMatrix)[2],
  quantiles = c(0.025, 0.05, 0.25, 0.5, 0.75, 0.95, 0.975)
)

Arguments

simpleTimeseries
- With a "Growth" estimate as a normally distributed quantility

yMatrix
- the matrix of possible infectivity profiles as discrete distributions

aVector
- the upper boundaries of the time cut-offs for the infectivity profiles

bootstraps
- the number of bootstraps to take to calculate for each point.

quantiles
- quantiles to calculate.

Value

a timeseries with "Rt" estimates